Changed nilled elements to return distinct value.
By default ZSI, when parsing SOAP packets, treats explicitly nilled
elements and missing optional elements the same way: it stores a None
value. For most uses this is adequate, but for our use we need to treat
these differently, as an explicit nil is interpreted as "clear the value
of this field" and an omitted value is interpreted as "do not change this
field if it is already defined".

The true irony is that internally ZSI is already prepared for this, as an
explicit nil is stored with the value of Nilled, which just happens to be
initialized to a Python None. We change this to a new ExplicitNil object
so now we can tell the difference in other code.

This does mean we now have two "null" values to test for (None and
ExplicitNil) but since we always know from the WSDL whether a particular
element can be nilled or not and whether it can be omitted or not, we know
when to make either test.